rename to animate_fates
